requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
matplotlib==3.8.4
seaborn==0.13.2
tqdm==4.66.4
scikit-learn==1.4.2
